net/mlx5e: Use size_t to store byte offset in statistics descriptors
The byte offset of counter descriptors should be stored in size_t variable
instead of an integer.

+++ drivers/net/ethernet/mellanox/mlx5/core/en_stats.h
@@ -48,7 +48,7 @@ FASTVC-BENCH-CORPUS:linux-main-100k-v1-e0bd41dc-8e12-4f1b-978d-a3645ae59da0
 
 struct counter_desc {
 	char		format[ETH_GSTRING_LEN];
-	int		offset; /* Byte offset */
+	size_t		offset; /* Byte offset */
 };
 
 struct mlx5e_sw_stats {
